@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:scoped_model/scoped_model.dart';
+import 'package:provider/provider.dart';
 import 'package:veggie_android_seasons/data/app_state.dart';
 import 'package:veggie_android_seasons/data/veggie.dart';
 import 'package:veggie_android_seasons/veggie_styles.dart';
@@ -8,7 +8,7 @@ import 'package:veggie_android_seasons/widgets/veggie_headline.dart';
 class FavouritesScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    final model = ScopedModel.of<AppState>(context, rebuildOnChange: true);
+    final model = Provider.of<AppState>(context);
     return Container(
       child: Center(
           child: model.favouriteVeggies.isEmpty
